Fix get external plugin config
